Refactor Flask demo application for Kubernetes deployment
- Updated `pyproject.toml` to include build system requirements and package finding configuration.
- Removed obsolete deployment configuration from `remediation.yaml`.
- Added Dockerfile for building and running the Flask application with Gunicorn.
- Created `incident_report.json` for logging application metrics and errors.
- Implemented Kubernetes deployment and service configuration in `k8s_deployment.yaml`.
- Developed main application logic in `main.py` with Prometheus metrics and error handling.
- Added README instructions for Azure and Kubernetes setup.
- Included requirements in `requirements.txt` for necessary dependencies.
- Enhanced `remediation.yaml` for improved deployment readiness and liveness probes.

+++ aiops_agent/analyzer/utlis.py
@@ -5,6 +5,6 @@ def _num(x: Any) -> Optional[float]:
     if x is None:
         return None
     try:
-        return float(x)
+        return int(x)
     except (TypeError, ValueError):
         return None
